Внесены правки. Посмотри пожалуйста на вопрос на 123 строчке в файле validation.php
