+++ templates/add-post/add-photo-post.php
@@ -1,7 +1,7 @@
 <div class="adding-post__input-wrapper form__input-wrapper">
   <label class="adding-post__label form__label" for="photo-url">Ссылка из интернета</label>
   <div class="form__input-section <?= (!empty($errors['photo-url'])) ? "form__input-section--error" : "" ?>">
-    <input class="adding-post__input form__input " id="photo-url" type="text" name="photo-url" placeholder="Введите ссылку" value="<?= getPostValue('photo-url') ?>">
+    <input class="adding-post__input form__input " id="photo-url" type="text" name="photo-url" placeholder="Введите ссылку" value="<?= anti_xss(getPostValue('photo-url')) ?>">
     <button class="form__error-button button" type="button">!<span class="visually-hidden">Информация об ошибке</span></button>
     <div class="form__error-text">
       <h3 class="form__error-title">Обнаружена ошибка</h3>
