+++ templates/add-post/add-text-post.php
@@ -1,7 +1,7 @@
 <div class="adding-post__textarea-wrapper form__textarea-wrapper">
   <label class="adding-post__label form__label" for="post-text">Текст поста <span class="form__input-required">*</span></label>
   <div class="form__input-section <?= (!empty($errors['post-text'])) ? "form__input-section--error" : "" ?>">
-    <textarea class="adding-post__textarea form__textarea form__input" id="post-text" name="post-text" placeholder="Введите текст публикации"><?= getPostValue('post-text') ?></textarea>
+    <textarea class="adding-post__textarea form__textarea form__input" id="post-text" name="post-text" placeholder="Введите текст публикации"><?= anti_xss(getPostValue('post-text')) ?></textarea>
     <button class="form__error-button button" type="button">!<span class="visually-hidden">Информация об ошибке</span></button>
     <div class="form__error-text">
       <h3 class="form__error-title">Обнаружена ошибка</h3>
