+++ templates/add-post/add-video-post.php
@@ -1,7 +1,7 @@
 <div class="adding-post__input-wrapper form__input-wrapper">
   <label class="adding-post__label form__label" for="video-url">Ссылка youtube <span class="form__input-required">*</span></label>
   <div class="form__input-section <?= (!empty($errors['video-url'])) ? "form__input-section--error" : "" ?>">
-    <input class="adding-post__input form__input" id="video-url" type="text" name="video-url" placeholder="Введите ссылку на видео" value="<?= getPostValue('video-url') ?>">
+    <input class="adding-post__input form__input" id="video-url" type="text" name="video-url" placeholder="Введите ссылку на видео" value="<?= anti_xss(getPostValue('video-url')) ?>">
     <button class="form__error-button button" type="button">!<span class="visually-hidden">Информация об ошибке</span></button>
     <div class="form__error-text">
       <h3 class="form__error-title">Обнаружена ошибка</h3>
